исправлены варнинги

+++ Game/Game/ColorBuffer.h
@@ -15,7 +15,7 @@ namespace colorBuffer
 	void BeginDraw();
 	void EndDraw();
 
-	void SetPixel(int x, int y, uint8_t r, uint8_t g, uint8_t b, uint8_t a = 255);
+	void SetPixel(uint16_t x, uint16_t y, uint8_t r, uint8_t g, uint8_t b, uint8_t a = 255);
 
 } // namespace colorBuffer
